feat: provide authorization gates for posts and comments

+++ app/Http/Controllers/CommentController.php
@@ -4,6 +4,7 @@ namespace App\Http\Controllers;
 
 use App\Models\Comment;
 use Illuminate\Http\Request;
+use Illuminate\Support\Facades\Gate;
 
 class CommentController extends Controller
 {
@@ -50,6 +51,9 @@ class CommentController extends Controller
      */
     public function destroy(string $id)
     {
+        if(Gate::denies("delete-comment", Comment::find($id))){
+            abort("403","You cannot delete this comment");
+        }
         Comment::destroy($id);
     }
 }
